refactor(Elife): Rename
