+++ src/themes/elife/lib/referencesFormatter.ts
